requirements.txt
python-dotenv==0.19.0
tqdm==4.62.1
tinkoff~=0.1.1
pandas~=1.4.1
currencyconverter==0.16.8